+++ src/gwproxy/gwproxy.c
@@ -933,7 +933,7 @@ static int expand_conn_slot(struct gwp_wrk *w)
 	struct gwp_ctx *ctx = w->ctx;
 
 	if (gcs->nr >= gcs->cap) {
-		size_t new_cap = gcs->cap ? gcs->cap * 2 : 16;
+		uint32_t new_cap = gcs->cap ? gcs->cap * 2 : 16;
 		struct gwp_conn_pair **new_pairs;
 
 		new_pairs = realloc(gcs->pairs, new_cap * sizeof(*new_pairs));
@@ -942,7 +942,7 @@ static int expand_conn_slot(struct gwp_wrk *w)
 
 		gcs->pairs = new_pairs;
 		gcs->cap = new_cap;
-		pr_dbg(&ctx->lh, "Increased connection slot capacity to %zu", gcs->cap);
+		pr_dbg(&ctx->lh, "Increased connection slot capacity to %u", gcs->cap);
 	}
 
 	return 0;
@@ -994,7 +994,7 @@ static int shrink_conn_slot(struct gwp_wrk *w)
 	struct gwp_conn_slot *gcs = &w->conn_slot;
 	struct gwp_conn_pair **new_pairs;
 	struct gwp_ctx *ctx = w->ctx;
-	size_t new_cap;
+	uint32_t new_cap;
 
 	if (!gcs->pairs)
 		return 0;
@@ -1018,7 +1018,7 @@ static int shrink_conn_slot(struct gwp_wrk *w)
 	}
 	gcs->pairs = new_pairs;
 	gcs->cap = new_cap;
-	pr_dbg(&ctx->lh, "Connection slot capacity shrunk to %zu", gcs->cap);
+	pr_dbg(&ctx->lh, "Connection slot capacity shrunk to %u", gcs->cap);
 	return 0;
 }
